requirements.txt
streamlit
openai
numpy
python-dotenv
faiss-cpu
rank_bm25
geopy
pydantic